requirements.txt
Django==5.2.7
djangorestframework==3.16.1
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5